+++ nistbeacon/nistbeaconvalue.py
@@ -1,5 +1,5 @@
 """
-Copyright 2015 Peter Urda
+Copyright 2015-2016 Peter Urda
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -17,11 +17,9 @@ limitations under the License.
 import binascii
 import hashlib
 import json
-import struct
 from random import Random
 from xml.etree import ElementTree
 
-import nistbeacon.constants as cn
 from nistbeacon.nistbeaconcrypto import NistBeaconCrypto
 
 
@@ -31,6 +29,29 @@ class NistBeaconValue(object):
     It has all the normal properties of a NIST beacon API call,
     but stored as a python object
     """
+
+    _xml_template = (
+        '<?xml version="1.0" encoding="UTF-8" standalone="yes"?>'
+        '<record xmlns="http://beacon.nist.gov/record/0.1/">'
+        '<version>{0}</version>'
+        '<frequency>{1}</frequency>'
+        '<timeStamp>{2}</timeStamp>'
+        '<seedValue>{3}</seedValue>'
+        '<previousOutputValue>{4}</previousOutputValue>'
+        '<signatureValue>{5}</signatureValue>'
+        '<outputValue>{6}</outputValue>'
+        '<statusCode>{7}</statusCode>'
+        '</record>'
+    )
+
+    _KEY_FREQUENCY = 'frequency'
+    _KEY_OUTPUT_VALUE = 'outputValue'
+    _KEY_PREVIOUS_OUTPUT_VALUE = 'previousOutputValue'
+    _KEY_SEED_VALUE = 'seedValue'
+    _KEY_SIGNATURE_VALUE = 'signatureValue'
+    _KEY_STATUS_CODE = 'statusCode'
+    _KEY_TIMESTAMP = 'timeStamp'
+    _KEY_VERSION = 'version'
 
     def __init__(
             self,
@@ -94,19 +115,19 @@ class NistBeaconValue(object):
         # Compute JSON, XML strings
         self._json = json.dumps(
             {
-                cn.NIST_KEY_VERSION: self.version,
-                cn.NIST_KEY_FREQUENCY: self.frequency,
-                cn.NIST_KEY_TIMESTAMP: self.timestamp,
-                cn.NIST_KEY_SEED_VALUE: self.seed_value,
-                cn.NIST_KEY_PREVIOUS_OUTPUT_VALUE: self.previous_output_value,
-                cn.NIST_KEY_SIGNATURE_VALUE: self.signature_value,
-                cn.NIST_KEY_OUTPUT_VALUE: self.output_value,
-                cn.NIST_KEY_STATUS_CODE: self.status_code,
+                self._KEY_VERSION: self.version,
+                self._KEY_FREQUENCY: self.frequency,
+                self._KEY_TIMESTAMP: self.timestamp,
+                self._KEY_SEED_VALUE: self.seed_value,
+                self._KEY_PREVIOUS_OUTPUT_VALUE: self.previous_output_value,
+                self._KEY_SIGNATURE_VALUE: self.signature_value,
+                self._KEY_OUTPUT_VALUE: self.output_value,
+                self._KEY_STATUS_CODE: self.status_code,
             },
             sort_keys=True,
         )
 
-        self._xml = cn.NIST_XML_TEMPLATE.format(
+        self._xml = self._xml_template.format(
             self.version,
             self.frequency,
             self.timestamp,
@@ -119,15 +140,12 @@ class NistBeaconValue(object):
 
         # Signature checking
         sha512_hash = NistBeaconCrypto.get_hash(
-            self.version.encode() +
-            struct.pack(
-                '>1I1Q64s64s1I',
-                self.frequency,
-                self.timestamp,
-                binascii.a2b_hex(self.seed_value),
-                binascii.a2b_hex(self.previous_output_value),
-                int(self.status_code)
-            )
+            self.version,
+            self.frequency,
+            self.timestamp,
+            self.seed_value,
+            self.previous_output_value,
+            self.status_code,
         )
 
         sig_check_result = NistBeaconCrypto.verify(
@@ -316,14 +334,14 @@ class NistBeaconValue(object):
         # Our required values are "must haves". This makes it simple
         # to verify we loaded everything out of JSON correctly.
         required_values = {
-            cn.NIST_KEY_FREQUENCY: None,
-            cn.NIST_KEY_OUTPUT_VALUE: None,
-            cn.NIST_KEY_PREVIOUS_OUTPUT_VALUE: None,
-            cn.NIST_KEY_SEED_VALUE: None,
-            cn.NIST_KEY_SIGNATURE_VALUE: None,
-            cn.NIST_KEY_STATUS_CODE: None,
-            cn.NIST_KEY_TIMESTAMP: None,
-            cn.NIST_KEY_VERSION: None,
+            cls._KEY_FREQUENCY: None,
+            cls._KEY_OUTPUT_VALUE: None,
+            cls._KEY_PREVIOUS_OUTPUT_VALUE: None,
+            cls._KEY_SEED_VALUE: None,
+            cls._KEY_SIGNATURE_VALUE: None,
+            cls._KEY_STATUS_CODE: None,
+            cls._KEY_TIMESTAMP: None,
+            cls._KEY_VERSION: None,
         }
 
         for key in required_values:
@@ -336,16 +354,16 @@ class NistBeaconValue(object):
 
         # We have all the required values, return a node object
         return cls(
-            version=required_values[cn.NIST_KEY_VERSION],
-            frequency=int(required_values[cn.NIST_KEY_FREQUENCY]),
-            timestamp=int(required_values[cn.NIST_KEY_TIMESTAMP]),
-            seed_value=required_values[cn.NIST_KEY_SEED_VALUE],
+            version=required_values[cls._KEY_VERSION],
+            frequency=int(required_values[cls._KEY_FREQUENCY]),
+            timestamp=int(required_values[cls._KEY_TIMESTAMP]),
+            seed_value=required_values[cls._KEY_SEED_VALUE],
             previous_output_value=required_values[
-                cn.NIST_KEY_PREVIOUS_OUTPUT_VALUE
+                cls._KEY_PREVIOUS_OUTPUT_VALUE
             ],
-            signature_value=required_values[cn.NIST_KEY_SIGNATURE_VALUE],
-            output_value=required_values[cn.NIST_KEY_OUTPUT_VALUE],
-            status_code=required_values[cn.NIST_KEY_STATUS_CODE],
+            signature_value=required_values[cls._KEY_SIGNATURE_VALUE],
+            output_value=required_values[cls._KEY_OUTPUT_VALUE],
+            status_code=required_values[cls._KEY_STATUS_CODE],
         )
 
     @classmethod
@@ -367,14 +385,14 @@ class NistBeaconValue(object):
         # Our required values are "must haves". This makes it simple
         # to verify we loaded everything out of XML correctly.
         required_values = {
-            cn.NIST_KEY_FREQUENCY: None,
-            cn.NIST_KEY_OUTPUT_VALUE: None,
-            cn.NIST_KEY_PREVIOUS_OUTPUT_VALUE: None,
-            cn.NIST_KEY_SEED_VALUE: None,
-            cn.NIST_KEY_SIGNATURE_VALUE: None,
-            cn.NIST_KEY_STATUS_CODE: None,
-            cn.NIST_KEY_TIMESTAMP: None,
-            cn.NIST_KEY_VERSION: None,
+            cls._KEY_FREQUENCY: None,
+            cls._KEY_OUTPUT_VALUE: None,
+            cls._KEY_PREVIOUS_OUTPUT_VALUE: None,
+            cls._KEY_SEED_VALUE: None,
+            cls._KEY_SIGNATURE_VALUE: None,
+            cls._KEY_STATUS_CODE: None,
+            cls._KEY_TIMESTAMP: None,
+            cls._KEY_VERSION: None,
         }
 
         # First attempt to load the xml, return 'None' on ParseError
@@ -403,14 +421,14 @@ class NistBeaconValue(object):
 
         # We have all the required values, return a node object
         return cls(
-            version=required_values[cn.NIST_KEY_VERSION],
-            frequency=int(required_values[cn.NIST_KEY_FREQUENCY]),
-            timestamp=int(required_values[cn.NIST_KEY_TIMESTAMP]),
-            seed_value=required_values[cn.NIST_KEY_SEED_VALUE],
+            version=required_values[cls._KEY_VERSION],
+            frequency=int(required_values[cls._KEY_FREQUENCY]),
+            timestamp=int(required_values[cls._KEY_TIMESTAMP]),
+            seed_value=required_values[cls._KEY_SEED_VALUE],
             previous_output_value=required_values[
-                cn.NIST_KEY_PREVIOUS_OUTPUT_VALUE
+                cls._KEY_PREVIOUS_OUTPUT_VALUE
             ],
-            signature_value=required_values[cn.NIST_KEY_SIGNATURE_VALUE],
-            output_value=required_values[cn.NIST_KEY_OUTPUT_VALUE],
-            status_code=required_values[cn.NIST_KEY_STATUS_CODE],
+            signature_value=required_values[cls._KEY_SIGNATURE_VALUE],
+            output_value=required_values[cls._KEY_OUTPUT_VALUE],
+            status_code=required_values[cls._KEY_STATUS_CODE],
         )